+++ vck-openid-ktor/src/commonTest/kotlin/at/asitplus/wallet/lib/ktor/openid/OpenId4VciClientExternalAuthorizationServerTest.kt
@@ -206,9 +206,6 @@ val OpenId4VciClientExternalAuthorizationServerTest by matrixSuite {
                     val params = requestBody.decodeFromPostBody<TokenIntrospectionRequest>()
                     externalAuthorizationServer.tokenIntrospection(
                         request = params,
-                        acceptHeader = request.headers.getAll(HttpHeaders.Accept)
-                            .shouldNotBeNull()
-                            .joinToString(", "),
                         httpRequest = request.toRequestInfo(),
                     ).fold(
                         onSuccess = { respond(it) },